+++ Sources/RichTextKit/Attributes/RichTextAttributeReader+Style.swift
@@ -13,7 +13,7 @@ public extension RichTextAttributeReader {
     /// Get the text styles at a certain range.
     func richTextStyles(at range: NSRange) -> [RichTextStyle] {
         let attributes = richTextAttributes(at: range)
-        let traits = font(at: range)?.fontDescriptor.symbolicTraits
+        let traits = richTextFont(at: range)?.fontDescriptor.symbolicTraits
         var styles = traits?.enabledRichTextStyles ?? []
         if attributes.isStrikethrough { styles.append(.strikethrough) }
         if attributes.isUnderlined { styles.append(.underlined) }