+++ Sources/RichTextKit/Attributes/RichTextAttributeWriter+Colors.swift
@@ -17,7 +17,7 @@ public extension RichTextAttributeWriter {
     ) {
         setRichTextAttribute(.backgroundColor, to: color, at: range)
     }
-
+    
     /// Set the foreground color at a certain range.
     func setRichTextForegroundColor(
         _ color: ColorRepresentable,
@@ -25,4 +25,29 @@ public extension RichTextAttributeWriter {
     ) {
         setRichTextAttribute(.foregroundColor, to: color, at: range)
     }
+    
+    /// Set the strikethrough color at a certain range.
+    func setRichTextStrikethroughColor(
+        _ color: ColorRepresentable,
+        at range: NSRange? = nil
+    ) {
+        setRichTextAttribute(.strikethroughColor, to: color, at: range)
+    }
+    
+    /// Set the stroke color at a certain range.
+    func setRichTextStrokeColor(
+        _ color: ColorRepresentable,
+        at range: NSRange? = nil
+    ) {
+        setRichTextAttribute(.strokeColor, to: color, at: range)
+    }
+    
+    /// Set the underline color at a certain range.
+    func setRichTextUnderlineColor(
+        _ color: ColorRepresentable,
+        at range: NSRange? = nil
+    ) {
+        setRichTextAttribute(.underlineColor, to: color, at: range)
+    }
+    
 }