+++ Sources/RichTextKit/Context/RichTextContext.swift
@@ -306,7 +306,7 @@ public extension RichTextContext {
     /// Set the attributed string to a new rich text.
     func setAttributedString(to string: NSAttributedString) {
         let mutable = NSMutableAttributedString(attributedString: string)
-        mutable.setFontSize(fontSize)
+        mutable.setRichTextFontSize(fontSize)
         shouldSetAttributedString = mutable
     }